// src/core/elements/editor/enums/symbol.md.enum.ts
export enum SymbolMdEnum {
    BOLD = '**',
    ITALIC = '_',
    LI = '- ',
    OL = '1. ',
    LI_CHECKED = '- [ ] ',
    H1 = '# ',
    H2 = '## ',
    H3 = '### ',
    H4 = '#### ',
    H5 = '##### ',
    H6 = '###### ',
    QUOTE = '> ',
    CODE = '```',
    CODE_INLINE = '`',
    LINK = '[title](http://)',
    IMAGE = '![alt](http://)',
    RESALT_QUOTE_START = '<span style="color: #4A9A6E">**',
    RESALT_AUTHOR_START = '<span style="color: #B079AB">**',
    END_SPAN = '**</span>',
}

export enum SymbolButtonEnum {
    BOLD = 'md-bold',
    ITALIC = 'md-italic',
    LI = 'md-li',
    OL = 'md-ol',
    LI_CHECKED = 'md-list-check',
    H1 = 'md-h1',
    H2 = 'md-h2',
    H3 = 'md-h3',
    H4 = 'md-h4',
    H5 = 'md-h5',
    H6 = 'md-h6',
    H = 'md-h',
    QUOTE = 'md-quote',
    CODE = 'md-code',
    CODE_INLINE = 'md-code-inline',
    RESALT_QUOTE = 'md-resalt-quote',
    RESALT_AUTHOR = 'md-resalt-author',
}